src/python/GatekeeperIntegrationAPI/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
tenacity==8.2.3
presidio-analyzer==2.2.351
presidio-anonymizer==2.2.351
azure-appconfiguration-provider==1.0.0
azure-identity==1.14.0
azure-keyvault-secrets==4.7.0
azure-monitor-opentelemetry==1.1.0
pylint==3.0.2
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry.instrumentation.logging==0.42b0
azure-monitor-opentelemetry-exporter==1.0.0b19
opentelemetry-exporter-jaeger==1.21.0
